:recycle: refactor: type 파일 basic, advanced 생성 및 기존 type 경로 수정, AdminPage 및 관련 컴포넌트에서 isAdmin 타입 사용, adminPage.new 잘 못 커밋 된 파일 삭제
