+++ src/advanced/components/CartPage.tsx
@@ -19,8 +19,8 @@ import { ProductList } from "./cart/ProductList";
 import { Cart } from "./cart/Cart";
 import { CouponSelector } from "./cart/CouponSelector";
 import { OrderSummary } from "./cart/OrderSummary";
-import { ProductWithUI } from "../../types";
-import { Coupon, CartItem } from "../../types";
+import { ProductWithUI } from "../type/types";
+import { Coupon, CartItem } from "../type/types";
 
 interface CartPageProps {
   isAdmin: boolean;
