+++ src/advanced/components/admin/ProductManagement.tsx
@@ -4,9 +4,10 @@ import { ProductTable } from "./ProductTable";
 import { ProductForm } from "./ProductForm";
 import { getDisplayPrice } from "../../utils/price";
 import { useProductForm } from "../../hooks/useProductForm";
-import { CartItem, Product, ProductWithUI } from "../../../types";
+import { CartItem, Product, ProductWithUI } from "../../type/types";
 
 interface ProductManagementProps {
+  isAdmin: boolean;
   products: ProductWithUI[];
   addProduct: (newProduct: Omit<ProductWithUI, "id">) => void;
   updateProduct: (productId: string, updates: Partial<ProductWithUI>) => void;
@@ -19,6 +20,7 @@ interface ProductManagementProps {
 }
 
 export function ProductManagement({
+  isAdmin,
   products,
   addProduct,
   updateProduct,
@@ -73,7 +75,7 @@ export function ProductManagement({
         onRemoveProduct={handleRemoveProduct}
         getRemainingStock={getRemainingStock}
         getDisplayPrice={getDisplayPrice}
-        isAdmin={true}
+        isAdmin={isAdmin}
       />
 
       <ProductForm
