+++ src/basic/components/cart/ProductList.tsx
@@ -1,5 +1,5 @@
 import { ProductImageIcon } from "../icons";
-import { ProductWithUI } from "../../../types";
+import { ProductWithUI } from "../../type/types";
 import { getDisplayPrice } from "../../utils/price";
 import { useDebounce } from "../../utils/hooks/useDebounce";
