+++ src/types.ts
@@ -3,12 +3,7 @@ export interface Product {
   name: string;
   price: number;
   stock: number;
-  description?: string;
   discounts: Discount[];
-}
-
-export interface ProductWithUI extends Product {
-  isRecommended?: boolean;
 }
 
 export interface Discount {
@@ -22,29 +17,6 @@ export interface CartItem {
 }
 
 export interface Coupon {
-  id?: string;
-  name: string;
-  code: string;
-  discountType: "amount" | "percentage";
-  discountValue: number;
-}
-
-export interface Notification {
-  id: string;
-  message: string;
-  type: "error" | "success" | "warning";
-}
-
-// 폼에서 사용할 타입들
-export interface ProductFormData {
-  name: string;
-  price: number;
-  stock: number;
-  description: string;
-  discounts: Discount[];
-}
-
-export interface CouponFormData {
   name: string;
   code: string;
   discountType: "amount" | "percentage";
